pyproject.toml
[project]
name = "g3m"
version = "0.0.1"
requires-python = ">= 3.10"

dependencies = [
    "ase>=3.24.0",
    "fire",
    "hydra-core>=1.3.0",
    "matplotlib==3.10.1",
    "numpy<2.0",
    "omegaconf==2.3.0",
    "pytorch-lightning==1.8.6",
    "rdkit==2023.9.5",
    "rich",
    "setuptools",
    "sympy>=1.11.1",
    "torch==2.1.2+cu121; sys_platform == 'linux'",
    "torch_cluster",
    "torch_geometric==2.5.2",
    "torch_scatter",
    "torch_sparse",
    "tqdm",
    "wandb==0.14.0",
]

[tool.setuptools.packages.find]
include = ["src_gmmm*"]

[tool.uv.sources]
torch = { index = "pytorch_linux", marker = "sys_platform == 'linux'" }

pyg-lib = [
    { url = "https://data.pyg.org/whl/torch-2.1.0%2Bcu121/pyg_lib-0.4.0%2Bpt21cu121-cp311-cp311-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
]
torch_cluster = [
    { url = "https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_cluster-1.6.3%2Bpt21cu121-cp311-cp311-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
]

torch_scatter = [
    { url = "https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_scatter-2.1.2%2Bpt21cu121-cp311-cp311-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
]

torch_sparse = [
    { url = "https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_sparse-0.6.18%2Bpt21cu121-cp311-cp311-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch_linux"
url = "https://download.pytorch.org/whl/cu121"
explicit = true